Convert testing init to a pytest fixture
